[plugin] Split ChiselComponent into its own file
